Update default ports to match dockerized id server (#76)
We now have support for dockerized identity server (which should also be
used for local development). They run on a different port than running
the identity server via vs studio. (We want an error if a developer runs
the identity server via vs studio instead of the docker image we
provide.

Co-authored-by: Mike Pullen <mike.pullen@swissgrc.com>