Add fbx generation (#107)
Add fbx generation with the help of blender.
To generate an fbx the server has to have blender installed (Which
blender executable is used can be defined via appsettings).

In the frontend the option to download fbx files has been added. This
will only work for estimations that have been created AFTER
this PR. Old files will throw an error when the user tries to download.
This is ok since we don't migrate old data.

Furthermore cleaned up the code so that it's easier to maintain and
expand. I decided to make QueueService an actual service. Previouslly it
was just a singleotn and didn't have an official scope set by ASP.NET.
Also changed the folder structucte so that each service has a separate
namespace and folder.
Also reduced duplicate code.